requirements.txt
aniso8601==9.0.1
black==23.7.0
blinker==1.6.2
click==8.1.7
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
Werkzeug==2.3.7
